add hacky hls.nix file
Motivation is to be able to sub-in our own GHC version, and I don't know
how to / if it's possible to do that when everything is defined inside a
flake. Ideally would have the flake use existing non-flake definitions.
Currently uses the .mwb suffix for haskell packages; should remove that
or abstract it.